Refactor: splitArrayInToChunks 함수 분리
